feat(STONEINTG-892): move metrics to pkg package
